Increase he precision of SOG and COG from float to double
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20808 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: dc869f5a5d03c1936007710b930cb0d5435aecf7

+++ server/wrapper/nmea/de.uniol.inf.is.odysseus.wrapper.nmea/src/de/uniol/inf/is/odysseus/wrapper/nmea/sentence/aissentences/payload/decoded/PositionReport.java
@@ -24,11 +24,11 @@ public abstract class PositionReport extends DecodedAISPayload {
 	
 	private final NavigationStatus navigationStatus;
 	private final Integer rateOfTurn;
-	private final Float speedOverGround;
+	private final Double speedOverGround;
 	private final Boolean positionAccurate;
 	private final Double latitude;
 	private final Double longitude;
-	private final Float courseOverGround;
+	private final Double courseOverGround;
 	private final Integer trueHeading;
 	private final Integer second;
 	private final ManeuverIndicator maneuverIndicator;
@@ -37,8 +37,8 @@ public abstract class PositionReport extends DecodedAISPayload {
 	protected PositionReport(String originalNmea, AISMessageType messageType,
 			Integer repeatIndicator, MMSI sourceMmsi,
 			NavigationStatus navigationStatus, Integer rateOfTurn,
-			Float speedOverGround, Boolean positionAccurate, Double latitude,
-			Double longitude, Float courseOverGround, Integer trueHeading,
+			Double speedOverGround, Boolean positionAccurate, Double latitude,
+			Double longitude, Double courseOverGround, Integer trueHeading,
 			Integer second, ManeuverIndicator maneuverIndicator,
 			Boolean raimFlag) {
 		super(originalNmea, messageType, repeatIndicator, sourceMmsi);
@@ -63,7 +63,7 @@ public abstract class PositionReport extends DecodedAISPayload {
 		return rateOfTurn;
 	}
 
-	public final Float getSpeedOverGround() {
+	public final Double getSpeedOverGround() {
 		return speedOverGround;
 	}
 
@@ -79,7 +79,7 @@ public abstract class PositionReport extends DecodedAISPayload {
 		return longitude;
 	}
 
-	public final Float getCourseOverGround() {
+	public final Double getCourseOverGround() {
 		return courseOverGround;
 	}
